+++ DAL.EF/DocumentScanEngine.cs
@@ -9,20 +9,24 @@ namespace DAL.EF
 {
 	public class DocumentScanEngine : IDocumentScanEngine
 	{
-		public DocumentScanEngine(ArchiveBookContext context)
+		public DocumentScanEngine()
 		{
-			this.context = context;
+			this.context = new ArchiveBookContext();
 		}
 
 		public List<DocumentScan> GetDocuments()
 		{
-			return context.DocumentScans.ToList();
+			List<DocumentScan> documents = context.DocumentScans.ToList();
+			//if (documents == null)
+			//	throw new ArgumentNullException("No documents");
+			//	return;
+			return documents;
 		}
 
 		public DocumentScan GetDocumentByID(int id)
 		{
 			if (id == 0)
-				throw new IndexOutOfRangeException("No such document.");
+				throw new ArgumentNullException("Invalid id.");
 
 			return context.DocumentScans.Where(d => d.ID_DocumentScan == id).FirstOrDefault();
 		}
@@ -30,7 +34,7 @@ namespace DAL.EF
 		public List<DocumentScan> GetDocumentsByArchiveBooking(int id)
 		{
 			if (id == 0)
-				throw new IndexOutOfRangeException("No such document.");
+				throw new ArgumentNullException("Invalid id.");
 
 			return context.DocumentScans.Where(d => d.ID_ArchiveBooking == id).ToList();
 		}
@@ -38,7 +42,9 @@ namespace DAL.EF
 		public void InsertDocument(DocumentScan document)
 		{
 			if (document == null)
-				throw new IndexOutOfRangeException("Empty document.");
+				throw new ArgumentNullException("Empty document.");
+
+			document.ID_DocumentScan = null;
 
 			context.DocumentScans.Add(document);
 			context.SaveChanges();
@@ -47,7 +53,10 @@ namespace DAL.EF
 		public void InsertDocuments(List<DocumentScan> documents)
 		{
 			if (documents.Count == 0)
-				throw new IndexOutOfRangeException("No documents.");
+				throw new ArgumentNullException("No documents.");
+
+			foreach (DocumentScan document in documents)
+				document.ID_DocumentScan = null;
 
 			context.DocumentScans.AddRange(documents);
 			context.SaveChanges();
@@ -56,9 +65,12 @@ namespace DAL.EF
 		public void DeleteDocumentById(int id)
 		{
 			if (id == 0)
-				throw new IndexOutOfRangeException("No such document.");
+				throw new ArgumentNullException("Invalid id.");
 
 			DocumentScan docScan = context.DocumentScans.Where(d => d.ID_DocumentScan == id).FirstOrDefault();
+			if (docScan == null)
+				//throw new ArgumentNullException("No document with that id in db.");
+				return;
 			context.DocumentScans.Remove(docScan);
 			context.SaveChanges();
 		}
@@ -66,9 +78,12 @@ namespace DAL.EF
 		public void DeleteDocumentByArchiveBooking(int id)
 		{
 			if (id == 0)
-				throw new IndexOutOfRangeException("No such document.");
+				throw new ArgumentNullException("Invalid id.");
 
 			DocumentScan docScan = context.DocumentScans.Where(d => d.ID_ArchiveBooking == id).FirstOrDefault();
+			if(docScan == null)
+				//throw new ArgumentNullException("No document from that archive booking in db.");
+				return;
 			context.DocumentScans.Remove(docScan);
 			context.SaveChanges();
 		}
@@ -76,12 +91,15 @@ namespace DAL.EF
 		public void UpdateDocumentScanByID(int id, DocumentScan document)
 		{
 			if (id == 0)
-				throw new IndexOutOfRangeException("No such document.");
+				throw new ArgumentNullException("Invalid id.");
 
 			if (document == null)
-				throw new IndexOutOfRangeException("Empty document.");
+				throw new ArgumentNullException("Empty document.");
 
 			DocumentScan docScan = context.DocumentScans.Where(d => d.ID_DocumentScan == id).FirstOrDefault();
+			if (docScan == null)
+				//throw new NullReferenceException("No document with that id in the db.");
+				return;
 			docScan.ID_ArchiveBooking = document.ID_ArchiveBooking;
 			docScan.DocumentPath = document.DocumentPath;
 			context.DocumentScans.Update(docScan);
